renaming modin_logging to logging
