+++ docs/usage_guide/advanced_usage/modin_logging.rst
@@ -5,6 +5,11 @@ Modin logging offers users greater insight into their queries by logging interna
 and profiling system memory. When Modin logging is enabled (default disabled), log files are written to a local `.modin` directory at the same
 directory level as the notebook/script used to run Modin. Each log file in the `.modin` directory is uniquely named after the job uuid. 
 
+**Developer Warning:** In some cases, running Modin logging inside of a Jupyterlab or other notebook instance at the root level of the Modin 
+repository or in the `modin/modin` directory may result in circular dependency issues. This is as a result of a name conflict between the 
+`modin/logging` directory and the Python `logging` module, which may be used as a default in such environments. As a resolution, one can simply
+run Modin logging from a different directory or manually manage the name conflicts.
+
 Usage example
 -------------
