+++ modin/core/execution/ray/implementations/pandas_on_ray/partitioning/partition.py
@@ -20,7 +20,7 @@ import uuid
 
 from modin.core.dataframe.pandas.partitioning.partition import PandasDataframePartition
 from modin.pandas.indexing import compute_sliced_len
-from modin.modin_logging import get_logger
+from modin.logging import get_logger
 
 ObjectIDType = ray.ObjectRef
 if version.parse(ray.__version__) >= version.parse("1.2.0"):
