+++ modin/core/io/text/text_file_dispatcher.py
@@ -35,7 +35,7 @@ from modin.utils import _inherit_docstrings
 from modin.core.io.text.utils import CustomNewlineIterator
 from modin.config import NPartitions
 from modin.error_message import ErrorMessage
-from modin.modin_logging.logger_function import logger_decorator
+from modin.logging import logger_decorator
 
 ColumnNamesTypes = Tuple[Union[pandas.Index, pandas.MultiIndex]]
 IndexColType = Union[int, str, bool, Sequence[int], Sequence[str], None]
