+++ modin/pandas/base.py
@@ -54,7 +54,7 @@ from modin.error_message import ErrorMessage
 import modin.pandas as pd
 from modin.pandas.utils import is_scalar
 from modin.config import IsExperimental
-from modin.modin_logging import LoggerMetaClass
+from modin.logging import LoggerMetaClass
 
 # Similar to pandas, sentinel value to use as kwarg in place of None when None has
 # special meaning and needs to be distinguished from a user explicitly passing None.
