+++ modin/pandas/indexing.py
@@ -36,7 +36,7 @@ from pandas.api.types import is_list_like, is_bool
 from pandas.core.dtypes.common import is_integer, is_bool_dtype, is_integer_dtype
 from pandas.core.indexing import IndexingError
 from modin.error_message import ErrorMessage
-from modin.modin_logging import LoggerMetaClass, metaclass_resolver
+from modin.logging import LoggerMetaClass, metaclass_resolver
 
 from .dataframe import DataFrame
 from .series import Series
